+++ alphadia/workflow/managers/raw_file_manager.py
@@ -5,9 +5,9 @@ import os
 
 import numpy as np
 
-from alphadia.data import DiaData
-from alphadia.data.alpharaw_wrapper import AlphaRawBase, MzML, Sciex, Thermo
-from alphadia.data.bruker import TimsTOFTranspose
+from alphadia.raw_data import DiaData
+from alphadia.raw_data.alpharaw_wrapper import AlphaRawBase, MzML, Sciex, Thermo
+from alphadia.raw_data.bruker import TimsTOFTranspose
 from alphadia.workflow.config import Config
 from alphadia.workflow.managers.base import BaseManager
