+++ alphadia/workflow/peptidecentric/extraction_handler.py
@@ -2,11 +2,11 @@ import pandas as pd
 import seaborn as sns
 from alphabase.spectral_library.base import SpecLibBase
 
-from alphadia.data import DiaData
 from alphadia.peakgroup import search
 from alphadia.peakgroup.config_df import HybridCandidateConfig
 from alphadia.plexscoring.config import CandidateConfig
 from alphadia.plexscoring.plexscoring import CandidateScoring
+from alphadia.raw_data import DiaData
 from alphadia.reporting.reporting import Pipeline
 from alphadia.workflow.config import Config
 from alphadia.workflow.managers.optimization_manager import OptimizationManager
