+++ alphadia/workflow/peptidecentric/fragment_requantification_handler.py
@@ -3,13 +3,13 @@ from alphabase.peptide.fragment import get_charged_frag_types
 from alphabase.spectral_library.base import SpecLibBase
 from alphabase.spectral_library.flat import SpecLibFlat
 
-from alphadia.data import DiaData
 from alphadia.fragcomp.utils import add_frag_start_stop_idx, candidate_hash
 from alphadia.plexscoring.config import CandidateConfig
 from alphadia.plexscoring.plexscoring import CandidateScoring
 from alphadia.plexscoring.utils import (
     candidate_features_to_candidates,
 )
+from alphadia.raw_data import DiaData
 from alphadia.reporting.reporting import Pipeline
 from alphadia.workflow.config import Config
 from alphadia.workflow.managers.calibration_manager import CalibrationManager
