+++ alphadia/workflow/peptidecentric/requantification_handler.py
@@ -1,13 +1,13 @@
 import pandas as pd
 from alphabase.spectral_library.base import SpecLibBase
 
-from alphadia.data import DiaData
 from alphadia.plexscoring.config import CandidateConfig
 from alphadia.plexscoring.plexscoring import CandidateScoring
 from alphadia.plexscoring.utils import (
     candidate_features_to_candidates,
     multiplex_candidates,
 )
+from alphadia.raw_data import DiaData
 from alphadia.reporting.reporting import Pipeline
 from alphadia.workflow.config import Config
 from alphadia.workflow.managers.calibration_manager import CalibrationManager
